pyproject.toml
[tool.poetry]
authors = ["Ken Van Haren <kenvanharen@gmail.com>"]
description = "Universal Data Schemas"
license = "BSD-3-Clause"
name = "common-model"
packages = [
  {include = "commonmodel"},
]
version = "0.3.1"

[tool.poetry.dependencies]
PyYAML = ">= 5.4.1"
pydantic = ">= 1.8.1"
python = ">= 3.7"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.1"
ipython = "^7.13.0"
isort = "^4.3.21"
mkdocs = "^1.1.2"
mkdocs-material = "^6.1.7"
mypy = "^0.770"
pre-commit = "^2.1.1"
pydeps = "^1.9.0"
pytest = "^4.6"
pytest-cov = "^2.8.1"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
